requirements.txt
fastapi
pydantic
httpx
fhir.resources
python-dotenv
requests
tabulate 